refactor: edit api /register to make register form to be able to upload profile pic and also add validation for full name and age
